# swagger.json
{
    "openapi": "3.0.0",
...
    "info": {
        "version": "1.0.0",
        "title": "API de Cortes de Luz",
        "description": "API utilizada para gestionar los datos de los domicilios, empresas y cortes de luz registrados"
    },
    "servers": [
        {
            "url": "http://localhost:2414"
        }
    ],
...
    "paths": {
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener una lista de todas las empresas",
                "description": "Obtiene una lista de empresas con sus respectivos datos",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "direccion": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Añade una nueva empresa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Avenida"
                                    }
                                },
                                "required": [
                                    "nombre",
                                    "direccion"
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/empresas/{nombre}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener una empresa buscandola por su ID",
                "description": "Obtiene una empresa con sus respectivos datos",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "direccion": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Borrar una empresa buscandola por su ID",
                "description": "Borra una empresa",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "direccion": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar todos los datos de una empresa",
                "description": "Actualiza los valores de una empresa",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nombre": {
                                        "required": true,
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "direccion": {
                                        "required": true,
                                        "type": "string",
                                        "example": "Avenida"
                                    }
                                },
                                "required": [
                                    "nombre",
                                    "direccion"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "direccion": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar algunos datos de una empresa",
                "description": "Actualiza ciertos valores de una empresa",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nombre": {
                                        "required": true,
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "direccion": {
                                        "required": true,
                                        "type": "string",
                                        "example": "Avenida"
                                    }
                                },
                                "required": [
                                    "nombre",
                                    "direccion"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "direccion": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}